requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
gspread==5.12.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
asyncio-throttle==1.0.2
pytest
pytest-asyncio
